SmallVector: Change the SizeType definition to avoid warnings
[Why]
If one does compile with -Wconversion enabled, one does get a lot of
warnings caused be SmallVector. Especially if one does use STL
functions like std::distance or std::forward etc...

+++ include/gul14/SmallVector.h
@@ -274,7 +274,7 @@ public:
     /// \copydoc ValueType
     using value_type = ValueType;
     /// Unsigned integer type for indexing, number of elements, capacity
-    using SizeType = uint32_t;
+    using SizeType = size_t;
     /// \copydoc SizeType
     using size_type = SizeType;
     /// Signed integer type for the difference of two iterators
